Reject invalid transitions, simplify transition table, require valid status in rpc
